Fix warnings
Turn on -Wall, -Wextra, and -Werror, and fix all of them

Marking LCM C++ bindings, lcm_coretypes.h, generated test data (because
it contains LCM C++ bindings) as system includes so we don't fail the
build on warnings in them (which do exist).

The build should both now be warning-free (except for third-party
code), and fail if new warnings appear.

Reviewers: hayk-skydio,bradley-solliday-skydio,nathan-skydio
Topic: sym-warnings

+++ symforce/opt/optimizer.h
@@ -239,7 +239,7 @@ class Optimizer {
    * Call nonlinear_solver_.Iterate on the given values (updating in place) until out of iterations
    * or converged
    */
-  void IterateToConvergence(Values<Scalar>* values, size_t num_iterations,
+  void IterateToConvergence(Values<Scalar>* values, int num_iterations,
                             bool populate_best_linearization, OptimizationStats<Scalar>* stats);
 
   /**
